server/pyproject.toml
[project]
name = "kenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi[standard]==0.115.12",
  "pydantic==2.11.3",
  "sqlmodel==0.0.24",
  "bcrypt==4.0.1",
  "passlib==1.7.4",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
]
